First overall draft - missing config files

+++ src/it/polimi/deib/atg/sharon/engine/ADLMatch.java
@@ -27,11 +27,11 @@ import java.util.ArrayList;
 public class ADLMatch {
 	private int HLadl;
 	private ArrayList<Integer> LLadl = new ArrayList<Integer>();
-	private ArrayList<Double> LLadlProbability = new ArrayList<Double>();
+	private ArrayList<Float> LLadlProbability = new ArrayList<Float>();
 	
 	
 	public ADLMatch(int hLadl, ArrayList<Integer> lLadl,
-                    ArrayList<Double> lLadlProbability) {
+                    ArrayList<Float> lLadlProbability) {
 		super();
 		HLadl = hLadl;
 		LLadl = lLadl;
@@ -59,12 +59,12 @@ public class ADLMatch {
 	}
 
 
-	public ArrayList<Double> getLLadlProbability() {
+	public ArrayList<Float> getLLadlProbability() {
 		return LLadlProbability;
 	}
 
 
-	public void setLLadlProbability(ArrayList<Double> lLadlProbability) {
+	public void setLLadlProbability(ArrayList<Float> lLadlProbability) {
 		LLadlProbability = lLadlProbability;
 	}
 }
